ai_platform/backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
pydantic>=2.7.2
sqlalchemy>=2.0.30
asyncpg>=0.29.0
celery>=5.4.0
redis>=5.0.4
httpx>=0.27.0
docker>=7.1.0
prometheus-fastapi-instrumentator>=7.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
alembic>=1.13.1
pydantic-settings>=2.2.1
